Modificacion Tipos de datos y datos nulos

+++ Base de datos/Customers.cs
@@ -30,7 +30,7 @@ namespace Base_de_datos
         public string firstName { get; set; }
         public string lastName { get; set; }
         public string identification { get; set; }
-        public int telephonNumber { get; set; }
+        public long telephonNumber { get; set; }
         public string cumulativeProfit { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
